convert page to ts
